Back out sign compare changes that appear to cause problems on hppa.
Back out sign compare changes that appear to cause problems on hppa.
Requested by deraadt@
Members:
	Makefile:1.36->1.37
	c_ksh.c:1.53->1.54
	edit.c:1.59->1.60
	emacs.c:1.78->1.79
	eval.c:1.55->1.56
	expand.h:1.13->1.14
	history.c:1.76->1.77
	lex.c:1.75->1.76
	lex.h:1.18->1.19
	misc.c:1.62->1.63
	path.c:1.20->1.21
	tree.c:1.28->1.29
	var.c:1.61->1.62
	vi.c:1.51->1.52

Date: 2018/01/04 20:06:16
Author: millert
Patchset: 745

+++ history.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: history.c,v 1.76 2018/01/01 19:45:56 millert Exp $	*/
+/*	$OpenBSD: history.c,v 1.77 2018/01/04 19:06:16 millert Exp $	*/
 
 /*
  * command history
@@ -552,7 +552,7 @@ sethistcontrol(const char *str)
 void
 sethistsize(int n)
 {
-	if (n > 0 && (uint32_t)n != histsize) {
+	if (n > 0 && n != histsize) {
 		int offset = histptr - history;
 
 		/* save most recent history */
